+++ public/pages/workflow_detail/component_input/index.ts
@@ -3,6 +3,5 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './workflow_inputs';
 export * from './input_fields';
 export * from './component_input';
