+++ public/pages/workflow_detail/component_input/ingest_inputs/index.ts
@@ -3,6 +3,5 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './ingest_inputs';
 export * from './ingest_data';
 export * from './source_data';
