+++ public/pages/workflow_detail/component_input/processor_inputs/ml_processor_inputs/modals/configure_template_modal.tsx
