+++ public/pages/workflow_detail/component_input/search_inputs/index.ts
@@ -3,5 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './search_inputs';
 export * from './configure_search_request';
