--- public/pages/workflow_detail/workflow_inputs/component_input/index.ts
@@ -1,6 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './component_input';
